requirements.txt
streamlit==0.58.0
face_recognition==1.3.0
Pillow==7.1.2
